Ensure slimmer doesn't mess with unicode characters.
I expect this test to fail on the continuous integration machine,
because of a bug in the version of libxml that Nokogiri is using on our
Linux boxes.

I want to be sure this test fails, before I commit a fix which works
around the problem. A better solution would be to upgrade
Nokogiri/libxml to include the relevant bug fix.