The ETag and Last-Modified are needed to allow conditional get requests.
I'm not convinced that slimmer is the correct place to filter headers (maybe the router would be better?).  For now though, it's important that both ETag and Last-Modified headers are allowed back to the browser, or conditional get requests will not work correctly.  The absence of these headers is also likely to adversely affect the performance of Varnish.

P.S. I've added some crude tests; not great, but better than the nothing that was there before.