Add some custom exceptions to slimmer
I've been meaning to get to this at the moment. Because of the way middleware exceptions are handled in rails apps it can be confusing when there's a slimmer error. I'm now capturing the two I've seen most commonly and raising custom exceptions that should make it easier to trace the issue quickly when it occurs