Allow replacement of `#wrapper` with arbitrary IDs.
The significant change here is that the output HTML will retain the ID of the source element, rather than retaining the `#wrapper` ID.

The content in #wrapper is inheriting a lot of CSS which is being brought in with the header and footer as part of the main template, but we don't want to have to build on top of that CSS in the future as this will cause dependency issues and complicate updating the header and footer itself.

By removing the `#wrapper` ID from the page, we can develop new CSS without having to override the existing styles with a stronger selector.