Allow configuration of custom wrapper IDs in middleware.
To do this, we needed to make the App configuration available within the Skin,
so we've refactored the arguments passed to be a Hash rather than the list of
parameters.

We also changed the arguments of the BodyInserter to be IDs rather than
selectors; they are converted to selectors internally. We did this because it
seems more natural to configure using IDs than CSS selector paths.