`request` is a misleading name for the original response headers
Not only is request a misleading name, I couldn't understand the
logic whereby the original headers were wrapped in a Rack::Request
object, and then extracted through the #env method.  As far as I
can tell, this is just an extremely complicated way to pass a hash
around.  As I want to use Rack::Response soon for my own purposes,
I've introduced it now as a replacement.