Copy conditional comments to template.
Conditional comments are the safest way of targeting specific IE
versions. As in inside government we are using SCSS to precompile our
CSS we can very easily create a desecrate version of stylesheets
targeting individual IEs.

After speaking with James Weiner he was in agreement that this would
also be a good way of stripping out our media queries to serve a
desktop sized stylesheet to IE while serving @media selectors to modern
browsers.

Added a new class to look at comments and find ones structured as an IE
conditional comment. Then move these accross to the template. Added a
test to check this.