Enable tags to be wrapped with IE conditionals
Being able to wrap tags with conditional comments will let us ensure we
don't serve modern stylesheets to old IE in the case where we are
serving them their own IE only stylesheet.

Added a proprietary attribute which takes the condition to add to the
comment in the form `slimmer-wrap-with="[condition]"`.

Added test to ensure the content that is wrapped is still visible by
normal browsers.