Remove meta viewport HTML if relevant header set
As an app that doesn't require a mobile view, remove the <meta viewport>
html that is in static. The main reason this commit exists is because
trade tariff doesn't need a mobile view.

However, the meta viewport html sets mobile devices so that they adjust
all content to fit their screen size and to turn off zoom. This allows
apps that don't want to enforce this the ability to turn it off using a
simple slimmer header.