Break out Slimmer::Skin#process_success to ease re-use
I want to use the success code without going through rack middleware. Splitting this method
out into one that does the rack processing and one that does the slimmer work enables that.
This then led to quite a few changes to processors to hand them more specific objects and less
rack responses.

There's more work to do to clean this up and to change to sane names. We may want to wait and do
that at the same time as deciding whether or not we want to use the old headers approach at all.
My feeling is that we don't as it won't work well with the new app but that's not for this commit.