Use rest-client to fetch templates
This will allow us to set extra headers more easily.

Note: this change also removes the option to skip verifying SSL certs.
This is a good thing.  We're using proper certs everywhere, so we should
be validating them.  This option being set is a holdover from the very
early days.