Return whether related section titles are to be shown
This method is made available to the related.raw.erb template, which is 
fetched from the Static app at runtime. It returns a value indicating 
whether section titles should be shown in the related items list.

It defaults to true, apart from when the header contains the value 
false (which gets cast as a string).