Adding BetaLabelInserter
A new Inserter Class for adding the Beta Label style. This will eventually
replace the Beta Notice as the de-facto way to show Users that a page or
service is in Beta.

To use this, send a selector with a position like so:
[Slimmer::Headers::BETA_LABEL] = "before:main.content"

We currently only accept before and after as the 2 options for where it
can be inserted.

This follows on from https://github.com/alphagov/slimmer/pull/64 and
https://github.com/alphagov/static/pull/381.