Include rack_env when notifying exceptions
We are seeing a number of exception notifications in whitehall that are
devoid of any environment information, including the request that caused
it. This has been traced back to the exception handling here in slimmer,
which notifies on exceptions in processors, but without including the
Rack environment. This change will mean the rack environment is passed
through to Airbrake such that it will be included in the notification
that gets generated.