Add HeaderIdentifier processor
The idea of this is to make every h1..h6 tag have an id so more savy
users can link to specific page sections.

This processor adds an id to any h* tag which doesn't already have one.
ID's are generated from the heading text prefixed with 'heading_'. Duplications
are avoided by appending '_2' or an iteration of when necessary.