Allow intergration tests to test components
So a network request isn't created in tests respond with a predefined
html block.

Create a test helper which generates a components selector for use in
a test framework.

The testing of the component and how it responds with different
arguments can be tested where the template exists. There is also a
further bit of work to create a intergration suit which will confirm
it runs correctly in preview/production.