Use production assets in test mode
Since preview was moved to its new home, we've seen a noticable drop in
performance. We suspect the slow loading of assets is what is causing
the sporadic failing of JS-enabled capybara scenarios on CI. This
updates the test templates to load assets from production instead of
preview.

The longer-term fix will be to get an instance of static running on CI
and load static assets directly from the local instance.