Move activesupport to a development dependancy
activesupport is used in the component resolver and is technically a
dependancy on the project.

However, the version of mocha we are using in this project doesn't work
with rails 4 gems. We therefore need to specify a rails 3 version of
activesupport so the tests will run. As this feature is only designed to
work in rails activesuport will always be loaded when this code isn't in
a test environment.