Don't swallow exceptions when running in test
Previously, exceptions from processors were always swallowed, and there
are several tests which trigger such exceptions but don't report them.
We want these exceptions to be reported in test and dev environments so
that we can fix them; continuing to swallow the exceptions in production
is probably a good thing.

Exceptions in production/preview are reported to errbit, as before.