Fix specified version of Plek
The version of Plek specified as the minimum requires an argument to be passed
to Plek.new, but we call it in component_resolver.rb without an argument.

Passing an argument was made optional in Plek 1.0.0, but 1.0.0 requires an
environment variable to be set. 1.1.0 and above don't, so lets specify that.