Support 410 responses
At the moment, if an app using slimmer returns a 410 response, the status code
is preserved, but the body is rewritten to the generic 500 error page.

Whilst pages which have been removed from GOV.UK should be registered as such
with the Router, if an app returns a 410, it should be honoured. Additionally,
if Slimmer is reused without the Router by third parties, then responses should
not be rewritten to look like server errors.

Static already has a 410 template, so this should "just work". I tested it
successfully in development.