Override #lookup in ComponentI18nBackup
The `#lookup` method is for loading the translation keys. By overriding
this method instead of `#translate` we retain the string interpolation
behaviour from the base i18n module.

Also refactored a few bits of code based on the observation that the
cache keys do not necessarily need to match the files or url paths
loaded from static, as long as they are consistent.

(paired with @edds on this)