Output rendering application in meta tag
It would be really useful to be able to segment analytics by rendering
application. It would let us have things like JavaScript errors per
application and page timings per application.

This change takes the class name of the currently running application
and adds it as a meta tag which can then be read by JavaScript.