Run GOV.UK ruby linter as part of CI builds
It's configured to only report on the diff of changes, not the whole repo.

I've run it against the whole repo and there a reasonable amount of violations,
and bulk fixing them is usually contentious, so i'm ignoring that for now.

With this linting in place at least nothing new will violate the style guide
and hopefully this will mean computers can give the feedback on style issues,
rather humans, which computers are better at and avoids humans having to give
nitpick-y feedback.