Add optional arg to template resolver call
CVE-2016-0752 was announced, the fix for which changed how Rails calls
template resolvers. It added an optional argument (defaulting to false)
which controls whether the template resolver should find templates
outside `Rails.root`. In Slimmer’s case, we take any requested
templates with a prefix of `govuk_component`and resolve them with the
static service over HTTP. The local file system is never touched.
Therefore, it’s safe to simply add the new argument here but ignore it
and carry on as normal.