Configure Jenkins build matrix
This introduces a matrix of builds under CI allowing us to test our gem
across a range of Ruby and Rails versions. This gives us confidence
that our gem will support a range of runtimes and host applications.

Our dependencies are not compatible across the full matrix so we need
to constrain them using different Gemfiles.