Add an ‘inside header’ inserter
This processor allows an application to inject a block of HTML after the logo by including a `.inside-header` element in their application’s output.

This is effectively like adding content to the inside_header yield block in govuk_template.

The service manual is using this to insert a title into the header.