Add support for Airbrake 5
At the moment, Slimmer attempts to call Airbrake.notify_or_ignore when
processing fails. The `notify_or_ignore` method was removed in Airbrake
5, which means that this fails in applications which use Slimmer and
Airbrake 5, which is most Rails 5 applications.

This adds a slightly more defensive check when attempting to notify
Airbrake, by confirming that `Airbrake` responds to `notify_or_ignore`.
If it doesn't, it checks to see if it responds to `notify`, and will
use that. If neither exist, it does nothing, as before.