Force request url to be UTF-8 encoding
This small change is the result of a long and complicated hunt into a
mysterious bug that was occurring approximately 4000 times a day on
Whitehall.

It is possible that GOV.UK may receive requests that have Unicode
characters in them. These are invalid as URIs as Unicode character must
be percent escaped. However unlike github - which will reject Unicode
characters with a 400 Bad Request - we will accept them.

This leaves is somewhat undefined waters where ruby servers may differ
in how they handle the request. Unicorn converts the request into an
ASCII encoded string. Normally this isn't a problem unless this string
contains a character outside of the range compatible with Unicode.

When this does occur though it raises an error when this string is
concatenated with the UTF-8 template. This can be resolved by forcing
the encoding to UTF-8.

This is done here as this will resolve this issue across all apps that
use Slimmer.

Alternatively this could be fixed in Whitehall by adding something like
this to Rack: https://github.com/whitequark/rack-utf8_sanitizer