Make inside header inserter more resilient
Some applications (like Service Manual Frontend) opt into using Slimmer for their javascript tests (so that they get e.g. jQuery) but in a test context the header logo is stubbed out, so this breaks.

This was previously masked by Slimmer swallowing exceptions raised within the processors, but this was removed in #203.