Make sure we insert meta tags above template metatags
When combining the template and application HTML, the `TagMover` will
take the meta and script tags from the application HTML and move them
to the bottom of the `<head>` tag.

`TagMover` moves them to the bottom, so that the CSS of the application
can override the CSS of the template (there's even a test in this gem
to make sure it does it like this).

However, the desired behaviour for meta tags is the reverse. In
particular, we want the `og:image` from the application to be higher
than the `og:image` from the template. This is because services like
Slack will choose the first `og:image` tag and not the more appropriate
custom one.

I tried to do this in
https://github.com/alphagov/govuk_template/pull/349, but I forgot that
life isn't easy and Slimmer exists.

Will enable https://github.com/alphagov/government-frontend/pull/878 to
work properly.