Drop cache TTL to 60 seconds
We experienced an [incident yesterday][1] that might have been caught
on staging if it wasn't for the delay between deploying to staging and
the caches expiring. By moving to a minute long cache we increase the
likelihood of noticing quickly after deploy.

The requests made are very fast anyway, so this is not likely to impact
load speed very much.

[1]
https://docs.google.com/document/d/1Hyg48BZVi0gVFkYO2FMDv7F797X3pOi-wbc7
pxzGTgE/edit#