Fix broken tests for initialized config
Previously we wrote tests that made use of an accidental behaviour
of the Rails middleware API. The API has rexently changed [1] to
use lambdas instead of array of arguments, and our tests were still
assuming the accidental output was an array of arrays. This fixes
the test to use the new data structure, and documents that this
approach is not ideal, in case someone decides to repeat it.

[1]: https://github.com/rails/rails/commit/672cc11f254ce3cb29aba2305c33d5f5e569e0e4#diff-533a9a9cc17a8a899cb830626089e5f9