Amend toggle button selector
The code in search_remover.rb assumes that the search toggle button is a
link.

This is an issue as work is being done to change the markup of this
button for accessibility purposes.

Change the search toggle button selector to a class selector instead to
ensure that functionality is preserved when we change the anchor tag to
a button tag.