Stop conflating all errors as "not found"
Previously we rescued any RestClient::Exception and then raised our
own TemplateNotFoundException, despite this being specific to a 404.
This updates the exception handling so that all non-specific errors
are correctly raised as a CouldNotRetrieveTemplate exception.

Note that RestClient does not provide classes for each kind of error;
instead, we need to inspect the exception to check the code.