Copy 'What Slimmer Does' documentation
At time of writing, this documentation (written in September 2012)
exists in the Slimmer Wiki on GitHub. We don't typically use wikis
in our repos, not least because we're beginning to import
documentation into https://docs.publishing.service.gov.uk/.
It contains useful context and should be more readily findable.

Our Rails conventions are that we store documentation in a folder
called `docs/` (https://github.com/alphagov/govuk-developer-docs/pull/2571),
and we plan on auto-ingesting docs folders into the Dev Docs in
the foreseeable future. It makes sense to have this useful doc
copied here in advance of that, and we can then delete the Wiki.

Note that Slimmer uses a thing called Yard to generate code
documentation, which was previously configured to output to a
'docs' directory, which was therefore git-ignored. It doesn't seem
to matter where the docs are saved to, so I've renamed the output
folder to `yard-docs` and removed `docs` from gitignore.