Ensure async and defer scripts are hoisted to the top of the head
This prevents `defer` or `async` script tags being automatically
moved to the bottom of the `body` as with other script tags.

Instead, these scripts will be hoisted to the top of the `head`,
above blocking tags such as CSS.

This will allow them to be fetched as soon as possible without
blocking parsing.