Add support for layout header component
During development the account nav markup was copied over from the
layout header component instead of using the component itself.
This was probably due to time constraints.

The component audits are complaining that we are including the assets
for layout_header, when we are not using the component.

To fix this problem, we are removing the hard coded account navigation
from static, and using the layout header component in its place. That
would technically be duplicating the header, so these changes are here
to ensure that the layout header is conditionally revealed only when the
account is enabled and the account nav should be present.