Remove navigation wrapper if empty
When the navigation links were removed in the gem layout it would leave the
wrapping element. The wrapping element had padding set on it which would cause
the layout to shift even if there were no child elements.

This change removes the wrapping element if there are no navigation links.

This is a `gem_layout` template specific change, so won't affect pages using the
`core_layout` template.