Use safe navigation operator in feedback url swapper
This was causing lots of tests to fail because the HTML stubs for the
tests didn't have the selector being searched for. This threw an error
because `set_attribute` can't be used on nil.

Using `&.` is equivalent to `thing ? thing.some_method : thing` - so
won't throw an error as `set_attribute` is not used.

Added a test to check that feedback_url_swapper can cope when a selector
is not present.