Resolve Nokogiri deprecation warning
Nokogiri warning:
warning: Passing a Node as the second parameter to Node.new is deprecated. Please pass a Document instead, or prefer an alternative constructor like Node#add_child. This will become an error in a future release of Nokogiri.

See 1.13.0 Nokogiri deprecation notice:
https://nokogiri.org/CHANGELOG.html#deprecated
"Passing a Nokogiri::XML::Node as the second parameter to Node.new is deprecated and will generate a warning. This parameter should be a kind of Nokogiri::XML::Document. This will become an error in a future version of Nokogiri. [#975]"
https://github.com/sparklemotion/nokogiri/issues/975