Decorate inline script elements with CSP nonce
We aspire [1] to use the GOV.UK Content Security Policy to only allow inline
script elements on GOV.UK with the use of a nonce attribute [2]. This is to
prevent cross-site-scripting in compliant browsers.

A challenge we have is that a rendering application (e.g. Frontend) may
have inline scripts and a template provided by static may also have
inline scripts and we need these to agree on the same nonce. We need to
add a nonce to an inline script in order to execute it. We do not need
one on scripts that have a src attribute as we expect that src to be
permitted by our script-src directive.

This change updates slimmer so that it will apply an applications nonce
within HTML returned by static. Static will not set any nonce
attributes.

For security, we need to run the NonceInserter processor before we
Slimmer inserts any HTML from an app. As otherwise, in a scenario where
someone found an ability to embed HTML in an app, this could mean we
inadvertently set a nonce on an XSS attempt. However if we run this
before any changes have been made to the template we can trust that it's
just as static provided it.

This approach depends on the availability of Rails (for the availability
of ActionDispatch::Request) which is an optional dependency of this gem.
I can't imagine a scenario currently where GOV.UK would be using this
nonce functionality without Rails so I've set the code to just fallback
quietly if used without Rails.

[1]: https://github.com/alphagov/govuk_app_config/blob/8680b784615339eadd6f716f35c7d0cad904bed1/lib/govuk_app_config/govuk_content_security_policy.rb#L55-L59
[2]: https://content-security-policy.com/nonce/